Flask-Transcription-App/Flask-Transcription-App/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-speech==2.30.0
google-cloud-translate==3.19.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.5
logging==0.4.9.6
MarkupSafe==3.0.2
marshmallow==3.26.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydub==0.25.1
pysrt==1.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.1
SQLAlchemy==2.0.40
standard-aifc==3.13.0
standard-chunk==3.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
